Update cli.py
New Feature:
Added a new flag (--full-compare) to enable comparison based on content of change_history table and not based on maximum version. In big projects this is easier to maintain because some changes take longer to be promoted and different objects have different versions.

Bugfix:
When a file is UTF8-BOM the Byte Order Marker is removed since Snowflake doesn't like this.