Add undo script functionality
Taking inspiration from Flyway, this commits adds the undo subcommand to
the application, with the `--step <n>` flag.

Undo scripts will have the `U` prefix, with the same name and version as
their corresponding versioned script. It will attempt to undo <n> amount
of versioned scripts until it finishes, or no corresponding undo script
is found.

Under the hood, the application is removing the `V` script from the
change history table, so when the `deploy` command is ran, the versioned
script can be applied again.

Algorithm:

* Fetch all applied `V` migrations
* Fetch all `U` migrations
* Get `U` migration of last `V` migration.
* IF `V` migration has no `U` migration, process stops
* Apply `U` script
* Delete `V` script from the change_history_table related to `U`

This commit closes issue #19